implement backend likes

+++ backend/routes/api/like.js
@@ -26,6 +26,8 @@ router.post('/', requireUser, async (req, res, next) =>{
             itinerary: req.body.itineraryId
         });
         await newLike.save();
+        await Itinerary.findByIdAndUpdate(req.body.itineraryId, { $inc: { likes: 1 } });
+        await User.findByIdAndUpdate(req.user._id, { $addToSet: { likes: req.body.itineraryId } })
         return res.status(201).json("liked")
     }catch(err){
         next(err)
@@ -36,11 +38,15 @@ router.delete('/:id', requireUser, async (req, res, next) =>{
     try {
         const like = await Like.findById(req.params.id);
         if (!like) {
-            res.status(404).send({ error: "Like doesn't exist!" });
-            return;
+            return res.status(404).send({ error: "Like doesn't exist!" });
+        }
+        if (like.user.toString() !== req.user._id.toString()) {
+            return res.status(403).json({ message: 'You do not have permission to delete this like' });
         }
 
         await Like.deleteOne({ _id: req.params.id });
+        await User.findByIdAndUpdate(req.user._id, { $pull: { likes: like.itinerary } });
+        await Itinerary.findByIdAndUpdate(like.itinerary, { $inc: { likes: -1 } })
         res.status(204).send();
     } catch (error) {
         res.status(500).send({ error: "An error occurred while deleting the like." });
